Fix project hierarchy and names
