+++ src/ipc-benchmark/boost/branch/performance.cpp
@@ -1,19 +1,19 @@
-#include <ipc-benchmark/boost/shm_branch.h>
+#include <ipc-benchmark/boost/branch/performance.h>
 
-#define __CLASS__ "[ShmBranch]"
+#define __CLASS__ "[Performance]"
 
 using namespace std::chrono_literals;
 
-namespace ipc_benchmark {
-	ShmBranch::ShmBranch() : Performance() {
+namespace ipc_benchmark::branch {
+	Performance::Performance() : ipc_benchmark::Performance() {
 		logger.trace(__CLASS__, __FUNCTION__);
 	}
 
-	ShmBranch::~ShmBranch() {
+	Performance::~Performance() {
 		logger.trace(__CLASS__, __FUNCTION__);
 	}
 
-	std::chrono::nanoseconds ShmBranch::run(uint64_t count) const {
+	std::chrono::nanoseconds Performance::run(uint64_t count) const {
 		logger.trace(__CLASS__, __FUNCTION__, count);
 		std::chrono::time_point begin = std::chrono::system_clock::now();
 		std::chrono::time_point end = std::chrono::system_clock::now();
