+++ src/ipc-benchmark/boost/branch/performance.h
@@ -1,0 +1,18 @@
+#ifndef __ipc_benchmark_boost_shm_branch_h__
+#define __ipc_benchmark_boost_shm_branch_h__
+
+#include <ipc-benchmark/base/performance.h>
+
+namespace ipc_benchmark {
+	namespace branch {
+		class Performance : public ipc_benchmark::Performance {
+		public:
+			Performance();
+			~Performance();
+		
+			std::chrono::nanoseconds run(uint64_t count) const override;
+		};
+	}
+}
+
+#endif
