+++ src/ipc-benchmark/boost/trunk/publisher.cpp
@@ -1,8 +1,8 @@
-#include <ipc-benchmark/boost/comm/publisher.h>
+#include <ipc-benchmark/boost/trunk/publisher.h>
 #include <ipc-benchmark/base/performance.h>
 #include <guutil/log/logger.h>
 
-namespace ipc_benchmark {
+namespace ipc_benchmark::trunk {
 	extern guutil::log::Logger logger; // global logger
 
 	Publisher::Publisher(std::string shmname, std::string topicname, uint64_t subscribers) {
